requirements.txt
pydub
SpeechRecognition
requests
python-telegram-bot
flask